+++ feedler/wagtail_hooks.py
@@ -3,9 +3,10 @@
 from wagtail.contrib.modeladmin.options import (
     ModelAdmin, modeladmin_register)
 
+from .admin import ExportModelAdminMixin
 from .models import Entry, Stream
 
-class EntryModelAdmin(ModelAdmin):
+class EntryModelAdmin(ExportModelAdminMixin, ModelAdmin):
     model = Entry
     menu_icon = 'date'
     menu_order = 200
